Odradjen izgled za formu
