+++ src/config/AxiosConfig.js
@@ -1,7 +1,7 @@
 import axios from "axios";
 
 const AxiosConfig = axios.create({
-  baseURL: "https://localhost:7251/", 
+  baseURL: "https://localhost:7251/",
 });
 
 export default AxiosConfig;
