+++ src/services/userServices.js
@@ -2,7 +2,10 @@ import AxiosConfig from "../config/axiosConfig";
 
 export const login = async (username, password) => {
   try {
-    const response = await AxiosConfig.post("/api/User/login", { username, password });
+    const response = await AxiosConfig.post("/api/User/login", {
+      username,
+      password,
+    });
     return response.data;
   } catch (error) {
     console.error("Login failed:", error.response?.data || error.message);
