+++ src/styles/global.scss
@@ -1,13 +1,118 @@
-body{
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
+$color-brown: #5d402b;
+$color-cream: #f5e4cb;
+$color-orange: #ea9332;
+body {
+  margin: 0;
+  padding: 0;
+  box-sizing: border-box;
 }
 
-.button-section{
+.button-section {
+  display: flex;
+  justify-content: flex-end;
+  gap: 20px;
+  background-color: $color-orange;
+  padding: 10px;
+}
+#LoginFormContainer {
+  width: 100%;
+  display: flex;
+  align-items: center;
+  flex-direction: column;
+  margin-top: 80px;
+
+  #LoginForm {
     display: flex;
-    justify-content: flex-end;
-    gap: 20px;
-    background-color: orange;
-    padding: 10px;
+    flex-direction: column;
+    align-items: center;
+    background-color: $color-cream;
+    box-shadow: 3px 3px 2px gray;
+    width: 350px;
+    border-radius: 10px;
+    padding: 20px;
+    .spinner {
+      border: 3px solid #f3f3f3; /* Light grey */
+      border-top: 3px solid #555; /* Darker */
+      border-radius: 50%;
+      width: 20px;
+      height: 20px;
+      animation: spin 0.8s linear infinite;
+      display: inline-block;
+      vertical-align: middle;
+      margin-left: 10px;
+    }
+
+    @keyframes spin {
+      0% {
+        transform: rotate(0deg);
+      }
+      100% {
+        transform: rotate(360deg);
+      }
+    }
+    input {
+      width: 250px;
+      height: 30px;
+      border-radius: 5px;
+      border-style: none;
+      margin: 5px;
+      padding: 10px;
+      font-size: 15px;
+      &::placeholder {
+        color: #000000;
+        opacity: 28%;
+      }
+    }
+    #password-container {
+      display: flex;
+      align-items: center;
+      flex-direction: row-reverse;
+      input {
+        width: 250px;
+        height: 30px;
+        border-radius: 5px;
+        border-style: none;
+        margin: 5px;
+        padding: 10px;
+        font-size: 15px;
+      }
+      button {
+        margin-right: 10px;
+        position: absolute;
+        width: 40px;
+        cursor: pointer;
+        background-color: white;
+        border: none;
+        margin-top: 5px;
+      }
+    }
+    #FormLoginInputsContainer {
+      display: block;
+    }
+    #LogInTitle {
+      text-align: center;
+      margin-top: 20px;
+      margin-bottom: 40px;
+    }
+    .LoginSubmitButton {
+      margin-top: 30px;
+      margin-bottom: 20px;
+      background-color: $color-brown;
+      color: white;
+      border: none;
+      padding: 15px;
+      padding-left: 20px;
+      padding-right: 20px;
+      border-radius: 5px;
+      cursor: pointer;
+      font-size: 15px;
+    }
+    .errorMessage {
+      color: $color-brown;
+      margin: 0;
+      margin-left: 10px;
+
+      font-size: 15px;
+    }
+  }
 }
